# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: HydroGuard water pipeline monitoring and demonstration API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/session:
    get:
      operationId: getAuthSession
      tags: [auth]
      summary: Get the signed-in user
      responses:
        "200":
          description: Current account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: No active session
    delete:
      operationId: logoutUser
      tags: [auth]
      summary: End the current session
      responses:
        "204":
          description: Session ended
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Account created and signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "400":
          description: Invalid registration details
        "409":
          description: Email or mobile number already registered
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Sign in with email or mobile and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Signed-in account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid credentials
...
  /maintenance:
    get:
      operationId: getMaintenance
      tags: [maintenance]
      summary: List maintenance records
      responses:
        "200":
          description: Repair records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaintenanceRecord"
    post:
      operationId: createMaintenance
      tags: [maintenance]
      summary: Create a maintenance record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenanceInput"
      responses:
        "201":
          description: Created maintenance record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceRecord"
  /maintenance/{id}:
    patch:
      operationId: updateMaintenance
      tags: [maintenance]
      summary: Update a maintenance record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenanceUpdate"
      responses:
        "200":
          description: Updated maintenance record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceRecord"
        "404":
          description: Maintenance record not found
...
  /profile:
    get:
      operationId: getProfile
      tags: [account]
      summary: Get current user's profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateProfile
      tags: [account]
      summary: Update current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /settings:
    get:
      operationId: getSettings
      tags: [account]
      summary: Get alert and device settings
      responses:
        "200":
          description: Current system settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemSettings"
    patch:
      operationId: updateSettings
      tags: [account]
      summary: Update alert settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Updated system settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemSettings"
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, fullName, email, mobile, role, assignedZone]
      properties:
        id: { type: string }
        fullName: { type: string }
        email: { type: string }
        mobile: { type: string }
        role: { type: string, enum: [administrator, water-champion] }
        assignedZone: { type: string }
    AuthSession:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
    RegisterInput:
      type: object
      required: [fullName, mobile, email, password, role]
      properties:
        fullName: { type: string, minLength: 2, maxLength: 120 }
        mobile: { type: string, minLength: 8, maxLength: 24 }
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        role: { type: string, enum: [administrator, water-champion] }
    LoginInput:
      type: object
      required: [identifier, password]
      properties:
        identifier: { type: string, minLength: 3 }
        password: { type: string, minLength: 1 }
...
    MaintenanceInput:
      type: object
      required: [pipeline, location, issue, priority]
      properties:
        pipeline: { type: string }
        location: { type: string }
        issue: { type: string }
        championId: { type: ["string", "null"] }
        priority: { type: string, enum: [low, medium, high, critical] }
        status: { type: string, enum: [pending, assigned, in-progress, completed] }
        alertId: { type: ["string", "null"] }
    MaintenanceUpdate:
      type: object
      properties:
        championId: { type: ["string", "null"] }
        priority: { type: string, enum: [low, medium, high, critical] }
        status: { type: string, enum: [pending, assigned, in-progress, completed] }
        issue: { type: string }
    MaintenanceRecord:
      type: object
      required: [id, maintenanceId, pipeline, location, issue, championName, priority, startedAt, completedAt, status]
      properties:
        id: { type: string }
        maintenanceId: { type: string }
        pipeline: { type: string }
        location: { type: string }
        issue: { type: string }
        championId: { type: ["string", "null"] }
        championName: { type: string }
        priority: { type: string, enum: [low, medium, high, critical] }
        startedAt: { type: string, format: date-time }
        completedAt: { type: ["string", "null"], format: date-time }
        status: { type: string, enum: [pending, assigned, in-progress, completed] }
        alertId: { type: ["string", "null"] }
...
    ProfileUpdate:
      type: object
      properties:
        fullName: { type: string, minLength: 2, maxLength: 120 }
        mobile: { type: string, minLength: 8, maxLength: 24 }
        email: { type: string, format: email }
        assignedZone: { type: string }
    SystemSettings:
      type: object
      required: [flowWarningThreshold, flowLeakThreshold, pressureWarningThreshold, pressureCriticalThreshold, devices]
      properties:
        flowWarningThreshold: { type: number }
        flowLeakThreshold: { type: number }
        pressureWarningThreshold: { type: number }
        pressureCriticalThreshold: { type: number }
        devices:
          type: array
          items:
            $ref: "#/components/schemas/Device"
    Device:
      type: object
      required: [id, name, location, status, lastSeen]
      properties:
        id: { type: string }
        name: { type: string }
        location: { type: string }
        status: { type: string, enum: [online, inactive] }
        lastSeen: { type: string, format: date-time }
        source: { type: string, enum: [field, simulator] }
        simulated: { type: boolean }
    SettingsUpdate:
      type: object
      properties:
        flowWarningThreshold: { type: number, minimum: 0 }
        flowLeakThreshold: { type: number, minimum: 0 }
        pressureWarningThreshold: { type: number, minimum: 0 }
        pressureCriticalThreshold: { type: number, minimum: 0 }