# lib/api-zod/src/generated/types/authSession.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * HydroGuard water pipeline monitoring and demonstration API
 * OpenAPI spec version: 0.2.0
 */
import type { UserProfile } from './userProfile';

export interface AuthSession {
  user: UserProfile;
}