# lib/api-zod/src/generated/types/maintenanceInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * HydroGuard water pipeline monitoring and demonstration API
 * OpenAPI spec version: 0.2.0
 */
import type { MaintenanceInputPriority } from './maintenanceInputPriority';
import type { MaintenanceInputStatus } from './maintenanceInputStatus';

export interface MaintenanceInput {
  pipeline: string;
  location: string;
  issue: string;
  /** @nullable */
  championId?: string | null;
  priority: MaintenanceInputPriority;
  status?: MaintenanceInputStatus;
  /** @nullable */
  alertId?: string | null;
}