# lib/api-zod/src/generated/types/maintenanceUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * HydroGuard water pipeline monitoring and demonstration API
 * OpenAPI spec version: 0.2.0
 */
import type { MaintenanceUpdatePriority } from './maintenanceUpdatePriority';
import type { MaintenanceUpdateStatus } from './maintenanceUpdateStatus';

export interface MaintenanceUpdate {
  /** @nullable */
  championId?: string | null;
  priority?: MaintenanceUpdatePriority;
  status?: MaintenanceUpdateStatus;
  issue?: string;
}