# lib/api-zod/src/generated/types/systemSettings.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * HydroGuard water pipeline monitoring and demonstration API
 * OpenAPI spec version: 0.2.0
 */
import type { Device } from './device';

export interface SystemSettings {
  flowWarningThreshold: number;
  flowLeakThreshold: number;
  pressureWarningThreshold: number;
  pressureCriticalThreshold: number;
  devices: Device[];
}